+++ harmony_service/__main__.py
@@ -1,19 +1,16 @@
-""" Run the Harmony Browse Image Generator Adapter via the Harmony CLI. """
+"""Run the Harmony Browse Image Generator Adapter via the Harmony CLI."""
 
 from argparse import ArgumentParser
 from sys import argv
 
 from harmony import is_harmony_cli, run_cli, setup_cli
 
-from harmony_browse_image_generator.adapter import BrowseImageGeneratorAdapter
-from harmony_browse_image_generator.exceptions import SERVICE_NAME
+from .adapter import BrowseImageGeneratorAdapter
+from .exceptions import SERVICE_NAME
 
 
 def main(arguments: list[str]):
-    """Parse command line arguments and invoke the appropriate method to
-    respond to them
-
-    """
+    """Parse command line arguments and invoke the appropriate method."""
     parser = ArgumentParser(
         prog=SERVICE_NAME, description='Run Harmony Browse Image Generator.'
     )