+++ tests/run_tests.sh
@@ -8,6 +8,10 @@
 # 2020-05-07: Adapted from SwotRepr project.
 # 2022-01-03: Removed safety checks, as these are now run in Snyk.
 # 2023-04-04: Updated for use with the Harmony Browse Image Generator (HyBIG).
+# 2024-07-30: Changes coverage to use pytest and output a unified
+# result. xmlrunner was unable to handle finding the tests in the separate
+# locations. Also use a relative path to the html output so that coverages can be
+# run outside of docker.
 #
 ###############################################################################
 
@@ -17,7 +21,7 @@ STATUS=0
 export HDF5_DISABLE_VERSION_CHECK=1
 
 # Run the standard set of unit tests, producing JUnit compatible output
-coverage run -m xmlrunner discover tests -o tests/reports
+coverage run -m pytest tests --junitxml=tests/reports/test-results-"$(date +'%Y%m%d%H%M%S')".xml
 RESULT=$?
 
 if [ "$RESULT" -ne "0" ]; then
@@ -28,14 +32,14 @@ fi
 echo "\n"
 echo "Test Coverage Estimates"
 coverage report --omit="tests/*"
-coverage html --omit="tests/*" -d /home/tests/coverage
+coverage html --omit="tests/*" -d tests/coverage
 
 # Run pylint
 # Ignored errors/warnings:
 # W1203 - use of f-strings in log statements. This warning is leftover from
 #         using ''.format() vs % notation. For more information, see:
 #     	  https://github.com/PyCQA/pylint/issues/2354#issuecomment-414526879
-pylint harmony_browse_image_generator --disable=W1203
+pylint hybig harmony_service --disable=W1203
 RESULT=$?
 RESULT=$((3 & $RESULT))