+++ tests/unit/test_sizes.py
@@ -10,9 +10,8 @@ from rasterio import Affine
 from rasterio.crs import CRS
 from rioxarray import open_rasterio
 
-from harmony_browse_image_generator.crs import PREFERRED_CRS
-from harmony_browse_image_generator.exceptions import HyBIGValueError
-from harmony_browse_image_generator.sizes import (
+from hybig.crs import PREFERRED_CRS
+from hybig.sizes import (
     METERS_PER_DEGREE,
     ScaleExtent,
     best_guess_target_dimensions,
@@ -120,7 +119,7 @@ class TestGetTargetGridParameters(TestCase):
         self.assertDictEqual(expected_parameters, actual_parameters)
 
     def test_grid_parameters_from_harmony_no_message_information(self):
-        """input granule is in preferred_crs on a 25km grid"""
+        """Input granule is in preferred_crs on a 25km grid"""
         crs = CRS.from_epsg(sp_seaice_grid['epsg'])
         height = sp_seaice_grid['height']
         width = sp_seaice_grid['width']
@@ -255,14 +254,14 @@ class TestTiling(TestCase):
             self.assertFalse(needs_tiling(grid_parameters))
 
     def test_get_cells_per_tile(self):
-        """test how tiles sizes are generated."""
+        """Test how tiles sizes are generated."""
         expected_cells_per_tile = self.CELLS_PER_TILE
         actual_cells_per_tile = get_cells_per_tile()
         self.assertEqual(expected_cells_per_tile, actual_cells_per_tile)
         self.assertIsInstance(actual_cells_per_tile, int)
 
     def test_compute_tile_boundaries_exact(self):
-        """tests subdivision of output image."""
+        """Tests subdivision of output image."""
         cells_per_tile = 10
         full_width = 10 * 4
         expected_origins = [0.0, 10.0, 20.0, 30.0, 40.0]
@@ -272,7 +271,7 @@ class TestTiling(TestCase):
         self.assertEqual(expected_origins, actual_origins)
 
     def test_compute_tile_boundaries_with_leftovers(self):
-        """tests subdivision of output image."""
+        """Tests subdivision of output image."""
         cells_per_tile = 10
         full_width = 10 * 4 + 3
         expected_origins = [0.0, 10.0, 20.0, 30.0, 40.0, 43.0]
@@ -282,7 +281,7 @@ class TestTiling(TestCase):
         self.assertEqual(expected_origins, actual_origins)
 
     def test_compute_tile_dimensions_uniform(self):
-        """test tile dimensions."""
+        """Test tile dimensions."""
         tile_origins = [0.0, 10.0, 20.0, 30.0, 40.0, 43.0]
         expected_dimensions = [10.0, 10.0, 10.0, 10.0, 3.0, 0.0]
 
@@ -291,7 +290,7 @@ class TestTiling(TestCase):
         self.assertEqual(expected_dimensions, actual_dimensions)
 
     def test_compute_tile_dimensions_nonuniform(self):
-        """test tile dimensions."""
+        """Test tile dimensions."""
         tile_origins = [0.0, 20.0, 35.0, 40.0, 43.0]
         expected_dimensions = [20.0, 15.0, 5.0, 3.0, 0.0]
 
@@ -299,8 +298,8 @@ class TestTiling(TestCase):
 
         self.assertEqual(expected_dimensions, actual_dimensions)
 
-    @patch('harmony_browse_image_generator.sizes.get_cells_per_tile')
-    @patch('harmony_browse_image_generator.sizes.needs_tiling')
+    @patch('hybig.sizes.get_cells_per_tile')
+    @patch('hybig.sizes.needs_tiling')
     def test_create_tile_output_parameters(
         self, needs_tiling_mock, cells_per_tile_mock
     ):
@@ -425,7 +424,6 @@ class TestChooseScaleExtent(TestCase):
 
     def test_scale_extent_from_input_image_and_no_crs_transformation(self):
         """Ensure no change of output extent when src_crs == target_crs"""
-
         with open_rasterio(
             self.fixtures / 'RGB.byte.small.tif', mode='r', mask_and_scale=True
         ) as in_array:
@@ -475,7 +473,7 @@ class TestChooseTargetDimensions(TestCase):
         actual_dimensions = choose_target_dimensions(message, None, scale_extent, None)
         self.assertDictEqual(expected_dimensions, actual_dimensions)
 
-    @patch('harmony_browse_image_generator.sizes.best_guess_target_dimensions')
+    @patch('hybig.sizes.best_guess_target_dimensions')
     def test_message_has_no_information(self, mock_best_guess_target_dimensions):
         """Test message with no information gets sent to best guess."""
         message = Message({})
@@ -489,7 +487,7 @@ class TestChooseTargetDimensions(TestCase):
             dataset, scale_extent, target_crs
         )
 
-    @patch('harmony_browse_image_generator.sizes.best_guess_target_dimensions')
+    @patch('hybig.sizes.best_guess_target_dimensions')
     def test_message_has_just_one_dimension(self, mock_best_guess_target_dimensions):
         """Message with only one dimension.