+++ CHANGELOG.md
@@ -4,6 +4,16 @@ HyBIG follows semantic versioning. All notable changes to this project will be
 documented in this file. The format is based on [Keep a
 Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [v2.0.0] - 2024-07-19
+
+**DAS-2180** - Adds pip installable library.
+
+This release is a refactor that extracts browse image generation logic from the
+harmony service code. There are no user visible changes to the existing
+functionality.  The new library,
+[hybig-py](https://pypi.org/project/hybig-py/), provides the `create_browse`
+function to generate browse images, see the README.md for details.
+
 ## [v1.2.2] - 2024-06-18
 
 ### Changed
@@ -52,7 +62,8 @@ outlined by the NASA open-source guidelines.
 For more information on internal releases prior to NASA open-source approval,
 see legacy-CHANGELOG.md.
 
-[unreleased]:https://github.com/nasa/harmony-browse-image-generator/compare/1.2.2..HEAD
+[unreleased]:https://github.com/nasa/harmony-browse-image-generator/compare/2.0.0..HEAD
+[v2.0.0]:https://github.com/nasa/harmony-browse-image-generator/compare/1.2.2..2.0.0
 [v1.2.2]: https://github.com/nasa/harmony-browse-image-generator/compare/1.2.1..1.2.2
 [v1.2.1]: https://github.com/nasa/harmony-browse-image-generator/compare/1.2.0..1.2.1
 [v1.2.0]: https://github.com/nasa/harmony-browse-image-generator/compare/1.1.0..1.2.0